* fix a bug in calculating accuracy when mask_zero is False

+++ tf2crf/crf.py
@@ -98,7 +98,7 @@ class CRF(tf.keras.layers.Layer):
         is_equal = K.equal(y_true, y_pred)
         is_equal = K.cast(is_equal, y_pred.dtype)
         if mask is None:
-            return K.sum(is_equal) / K.sum(self.sequence_lengths)
+            return K.sum(is_equal) / K.cast(K.sum(self.sequence_lengths), y_pred.dtype)
         else:
             mask = K.cast(mask, y_pred.dtype)
             return K.sum(is_equal * mask) / K.sum(mask)
